+++ build.gradle
@@ -32,6 +32,8 @@ dependencies {
 	implementation 'org.postgresql:postgresql:42.2.5'
 	implementation 'org.liquibase:liquibase-core:4.15.0' // Aggiungi Liquibase
 	implementation 'org.springframework.boot:spring-boot-starter-jdbc' // Per il supporto JDBC
+	implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
+
 
 
 	implementation 'org.projectlombok:lombok:1.18.30'
