+++ src/main/java/backend/model/Prodotto.java
@@ -1,0 +1,21 @@
+package backend.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.UUID;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "prodotto")
+public class Prodotto {
+    @Id
+    @Column(name = "id", nullable = false)
+    private UUID id;
+
+}
