+++ src/main/java/backend/model/Recensione.java
@@ -1,0 +1,36 @@
+package backend.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.GenericGenerator;
+
+import java.time.LocalDate;
+import java.util.UUID;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "recensione")
+public class Recensione {
+
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
+    @Column(name = "id", updatable = false, nullable = false)
+    private UUID id;
+
+    @Column(nullable = false)
+    private LocalDate data;
+
+    @Column(nullable = false)
+    private int numStelle;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Prodotto prodotto;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Utente utente;
+}
