--- src/main/java/it/unife/sample/backend/BackendApplication.java
@@ -1,13 +1,0 @@
-package it.unife.sample.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendApplication.class, args);
-	}
-
-}
