--- src/main/java/it/unife/sample/backend/model/SampleEntity.java
@@ -1,20 +1,0 @@
-package it.unife.sample.backend.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-@Entity
-public class SampleEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private UUID id;
-
-    private String name;
-}
