+++ src/main/resources/db/changelog/migration/Creazione-Recensione-VUtente-VProdotto.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.15.xsd"
+        objectQuotingStrategy="QUOTE_ONLY_RESERVED_WORDS">
+    <changeSet id="1745429461738-1" author="riccardo">
+        <createTable tableName="prodotto">
+            <column name="id" type="UUID">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_prodotto"/>
+            </column>
+        </createTable>
+    </changeSet>
+    <changeSet id="1745429461738-2" author="riccardo">
+        <createTable tableName="recensione">
+            <column name="id" type="UUID">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_recensione"/>
+            </column>
+            <column name="data" type="DATE">
+                <constraints nullable="false"/>
+            </column>
+            <column name="num_stelle" type="INT">
+                <constraints nullable="false"/>
+            </column>
+            <column name="prodotto_id" type="UUID">
+                <constraints nullable="false"/>
+            </column>
+            <column name="utente_id" type="UUID">
+                <constraints nullable="false"/>
+            </column>
+        </createTable>
+    </changeSet>
+    <changeSet id="1745429461738-3" author="riccardo">
+        <createTable tableName="utente">
+            <column name="id" type="UUID">
+                <constraints nullable="false" primaryKey="true" primaryKeyName="pk_utente"/>
+            </column>
+        </createTable>
+    </changeSet>
+    <changeSet id="1745429461738-4" author="riccardo">
+        <addForeignKeyConstraint baseColumnNames="prodotto_id" baseTableName="recensione"
+                                 constraintName="FK_RECENSIONE_ON_PRODOTTO" referencedColumnNames="id"
+                                 referencedTableName="prodotto"/>
+    </changeSet>
+    <changeSet id="1745429461738-5" author="riccardo">
+        <addForeignKeyConstraint baseColumnNames="utente_id" baseTableName="recensione"
+                                 constraintName="FK_RECENSIONE_ON_UTENTE" referencedColumnNames="id"
+                                 referencedTableName="utente"/>
+    </changeSet>
+
+</databaseChangeLog>
