--- src/test/java/it/unife/sample/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package it.unife.sample.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
